Renaming components to remove CanvasManifest wording
