+++ .notes
@@ -1,0 +1,2 @@
+sudo docker run -p 5984:5984 -d couchdb
+delete /couchdb/kivik.tmp
