+++ services/admin/src/components/manifests/ManifestCanvasSelector.svelte
