+++ services/admin/src/components/manifests/ManifestContentEditor.svelte
