+++ services/admin/src/components/manifests/ManifestEditorActions.svelte
