+++ services/admin/src/components/manifests/ManifestInfoEditor.svelte
