+++ services/admin/src/components/manifests/ManifestStatusIndicator.svelte
